evaluation/requirements.txt
# Azure AI Evaluation SDK
azure-ai-evaluation>=1.0.0

# Azure Identity (for credential-based auth)
azure-identity>=1.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP client for API calls
httpx>=0.25.0
aiohttp>=3.9.0

# Environment variable handling
python-dotenv>=1.0.0

# JSON Lines handling
jsonlines>=4.0.0
